mostly filled in result dao
Former-commit-id: c39b49ab31a9daa5cc52cf3f60d3c4b9a231fdb8 [formerly 51f7977804661b94745b4fd277038b06e2e78ae3]
Former-commit-id: 129474dd401caf2bb2cc98218c21876aeeb4cf3d

+++ src/mitll/langtest/client/scoring/ASRScoringAudioPanel.java
@@ -109,7 +109,7 @@ public class ASRScoringAudioPanel<T extends Shell> extends ScoringAudioPanel<T> 
    * @param height of images returned
    * @param reqid so if many requests are made quickly and the returns are out of order, we can ignore older requests
    */
-  protected void scoreAudio(final String path, long resultID, String refSentence,
+  protected void scoreAudio(final String path, int resultID, String refSentence,
                             final ImageAndCheck wordTranscript, final ImageAndCheck phoneTranscript,
                             int toUse, int height, final int reqid) {
     if (path == null) return;
